requirements.txt
django
djangorestframework
pillow
django-environ
djoser
djangorestframework-simplejwt
